Rewrite timelines - split them to file per prefix
Timeline generation needed lots and lots of memory. This is not acceptable
so new timeline format has been devised and it is easier to generate
it (but will see how it performs on real data, because there is going to
be much higher rate of I/O operations on files).